Add insert_method prop to block insertion track events
Takes a guess at which block insertion/replacement method was used by
inspecting the current state of Redux and the DOM when the
`insertBlock(s)` and `replaceBlock(s)` actions are fired on
`core/block-editor`.

Adds the `insert_method` prop to two existing tracks events:
- `wpcom_block_inserted`, when a block is inserted
- `wpcom_block_picker_block_inserted`, when a block is replaced

The prop can take the following values:
- `header-inserter`
- `slash-inserter`
- `quick-inserter`
- `block-inserter`
- or undefined

Undefined is used when we don't know what method was used. It could have
been the page layout picker, or it could have been pressing "return" at
the end of a paragraph block.