Fix ternary logic in PlanFeaturesActionsButton
I think the intent was to say "use the buttonText from props, or
determine the button text based on if the plan is free", but the first
two conditions were combined so instead the behavior was "if there is
buttonText in the props or if the plan is free, display the free plan
text and otherwise display the upgrade text".

This separates the two conditions so it should be more clear what
happens when.